@@ -8,6 +8,7 @@
     "serve:debug": "ng server --proxy-config proxy.conf.json  & node-debug server/index.js",
     "ng": "ng serve",
     "start": "nodemon server/index.js",
+    "dev": "ng serve --proxy-config proxy.conf.json",
     "build": "ng build --prod --aot=false",
     "lint": "tslint \"src/**/*.ts\" --project src/tsconfig.json --type-check && tslint \"e2e/**/*.ts\" --project e2e/tsconfig.json --type-check",
     "test": "ng test",
@@ -17,62 +18,64 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.11",
-    "@angular/common": "^5.2.11",
-    "@angular/compiler": "^5.2.11",
-    "@angular/core": "^5.2.11",
-    "@angular/forms": "^5.2.11",
-    "@angular/http": "^5.2.11",
-    "@angular/platform-browser": "^5.2.11",
-    "@angular/platform-browser-dynamic": "^5.2.11",
-    "@angular/platform-server": "^5.2.11",
-    "@angular/router": "^5.2.11",
-    "admin-lte": "^2.4.18",
-    "angular-dual-listbox": "^4.7.0",
+    "@angular/animations": "^9.1.0",
+    "@angular/common": "^9.1.0",
+    "@angular/compiler": "^9.1.0",
+    "@angular/core": "^9.1.0",
+    "@angular/forms": "^9.1.0",
+    "@angular/platform-browser": "^9.1.0",
+    "@angular/platform-browser-dynamic": "^9.1.0",
+    "@angular/platform-server": "^9.1.0",
+    "@angular/router": "^9.1.0",
+    "admin-lte": "^3.0.4",
+    "angular-dual-listbox": "^5.0.1",
     "animate.css": "^3.7.2",
-    "bootstrap": "^3.4.1",
+    "bootstrap": "^4.4.1",
     "chart.js": "^2.9.3",
-    "core-js": "^2.6.11",
-    "dotenv": "^4.0.0",
+    "core-js": "^3.6.4",
+    "dotenv": "^8.2.0",
     "font-awesome": "^4.7.0",
-    "glue": "^4.1.0",
-    "h2o2": "^5.4.0",
-    "inert": "^4.1.0",
+    "glue": "^6.0.0",
+    "h2o2": "^8.2.0",
+    "inert": "^5.1.3",
     "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
     "mydaterangepicker": "^4.2.1",
-    "ng2-charts": "^1.5.0",
+    "ng2-charts": "^2.3.0",
     "ng2-dual-list-box": "^1.0.5",
-    "ng2-slim-loading-bar": "^2.3.0",
-    "ng2-toasty": "^2.3.0",
-    "ngx-bootstrap": "~1.8.0",
+    "ng2-slim-loading-bar": "^4.0.0",
+    "ng2-toasty": "^4.0.3",
+    "ngx-bootstrap": "~5.5.0",
     "nodejslogger": "^1.2.3",
     "q": "^1.5.1",
-    "rxjs": "^5.5.12",
+    "rxjs": "^6.5.4",
+    "rxjs-compat": "^6.5.4",
     "ts-helpers": "^1.1.1",
-    "wreck": "^10.0.0",
+    "tslib": "^1.10.0",
+    "wreck": "^14.2.0",
     "xml2js": "^0.4.23",
-    "zone.js": "0.8.29"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.11",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "^6.14.9",
-    "codelyzer": "~2.0.0-beta.1",
-    "eslint": "^4.19.1",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "2.5.0",
-    "karma": "1.2.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.1.2",
-    "karma-remap-istanbul": "^0.2.1",
+    "@angular-devkit/build-angular": "~0.901.0",
+    "@angular/cli": "^9.1.0",
+    "@angular/compiler-cli": "^9.1.0",
+    "@types/jasmine": "3.5.10",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "eslint": "^6.8.0",
+    "jasmine-core": "3.5.0",
+    "jasmine-spec-reporter": "5.0.1",
+    "karma": "4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-cli": "^2.0.0",
+    "karma-jasmine": "^3.1.1",
+    "karma-remap-istanbul": "^0.6.0",
     "node-sass": "^4.13.1",
-    "protractor": "~4.0.13",
-    "rimraf": "^2.7.1",
-    "ts-node": "1.2.1",
-    "tslint": "^4.3.0",
-    "typescript": "^2.4.2"
+    "protractor": "~5.4.3",
+    "rimraf": "^3.0.2",
+    "ts-node": "8.8.1",
+    "tslint": "^6.1.1",
+    "typescript": "^3.8.3"
   }
 }
