@@ -6,15 +6,12 @@ import {LoginResponse} from './commons/models/common-data-models';
 @Component({
     selector: 'body',
     templateUrl: './app.component.html',
-    styles: [`:host {
-        background-color: blue
-    }`],
     styleUrls: ['./app.component.scss']
 })
 
 export class AppComponent implements OnInit {
-    private isLoggedIn = false;
-    private isMenuExpanded = false;
+    public isLoggedIn = false;
+    public isMenuExpanded = false;
 
     constructor(private _authenticationService: AuthenticationService,
                 private _appCommonService: AppCommonService) {
