@@ -1,24 +1,24 @@
-import {Configuration} from '../models/common-data-models';
-import {Injectable} from '@angular/core';
-import {Http} from '@angular/http';
+import { Configuration } from '../models/common-data-models';
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
 
 
 @Injectable()
 export class ConfigService {
     private config: Configuration;
 
-    constructor(private http: Http) {
+    constructor(private http: HttpClient) {
     }
 
     load(url: string) {
         return new Promise(
             (resolve) => {
-            this.http.get(url)
-                .map(res => res.json())
-                .subscribe(config => {
-                    this.config = config;
-                    resolve();
-                }); });
+                this.http.get<Configuration>(url)
+                    .subscribe(config => {
+                        this.config = config;
+                        resolve();
+                    });
+            });
     }
 
     getConfiguration(): Configuration {
