@@ -1,10 +1,12 @@
-import {Injectable} from '@angular/core';
-import {Headers, Http, RequestOptions, Response} from '@angular/http';
-import {BehaviorSubject, Observable, Subject} from 'rxjs';
-import {ApplicationTaskResult, ApplicationTaskResults} from '../commons/models/application-data-models';
-import {AuthenticationService} from '../commons/services/authentication.service';
-import {SlimLoadingBarService} from 'ng2-slim-loading-bar';
-import {MessageService} from '../commons/services/message.service';
+
+import { throwError as observableThrowError, BehaviorSubject, Observable, Subject } from 'rxjs';
+import { Injectable } from '@angular/core';
+import { ApplicationTaskResult, ApplicationTaskResults } from '../commons/models/application-data-models';
+import { AuthenticationService } from '../commons/services/authentication.service';
+import { SlimLoadingBarService } from 'ng2-slim-loading-bar';
+import { MessageService } from '../commons/services/message.service';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { catchError } from 'rxjs/operators';
 
 @Injectable()
 export class ApplicationRemoteDataService {
@@ -24,28 +26,29 @@ export class ApplicationRemoteDataService {
         search: this.apiContext + 'applications/search',
     };
 
-    constructor(private http: Http,
-                private slimLoadingBarService: SlimLoadingBarService,
-                private authService: AuthenticationService,
-                private message: MessageService) {
+    constructor(private http: HttpClient,
+        private slimLoadingBarService: SlimLoadingBarService,
+        private authService: AuthenticationService,
+        private message: MessageService) {
     }
 
     getMyApplicationTasks(): void {
         this.slimLoadingBarService.start();
         this.http.get(this.apiEndpoints['search'], this.getOptions())
-            .map((response: Response) => response.json())
-            .catch((error: Response) => Observable.throw({
-                success: false,
-                message: 'Error Loading My Application List',
-                error: error
-            }))
+            .pipe(
+                catchError((error: Response) => observableThrowError({
+                    success: false,
+                    message: 'Error Loading My Application List',
+                    error: error
+                }))
+            )
             .subscribe(
                 data => {
-                    if (data.success) {
-                        this.MyApplicationApprovalTasksProvider.next(data.payload);
+                    if (data['success']) {
+                        this.MyApplicationApprovalTasksProvider.next(data['payload']);
                         this.slimLoadingBarService.complete();
                     } else {
-                        this.message.error(data.message);
+                        this.message.error(data['message']);
                         this.slimLoadingBarService.stop();
                     }
                 },
@@ -58,40 +61,33 @@ export class ApplicationRemoteDataService {
 
     getAllApplicationTasks(): void {
         this.slimLoadingBarService.start();
-        this.http.get(this.apiEndpoints['search'], this.getOptions())
-            .map((response: Response) => response.json())
-            .catch((error: Response) => Observable.throw({
-                success: false,
-                message: 'Error Loading All Application List',
-                error: error
-            }))
-            .subscribe(
-                data => {
-                    if (data.success) {
-                        this.AllApplicationApprovalTasksProvider.next(data.payload);
-                        this.slimLoadingBarService.complete();
-                    } else {
-                        this.message.error(data.message);
-                        this.slimLoadingBarService.stop();
-                    }
-                },
-                error => {
-                    this.message.error(error.message);
+        this.http.get(this.apiEndpoints['search'], this.getOptions()).subscribe(
+            data => {
+                if (data['success']) {
+                    this.AllApplicationApprovalTasksProvider.next(data['payload']);
+                    this.slimLoadingBarService.complete();
+                } else {
+                    this.message.error(data['message']);
                     this.slimLoadingBarService.stop();
                 }
-            );
+            },
+            error => {
+                this.message.error(error.message);
+                this.slimLoadingBarService.stop();
+            }
+        );
     }
 
-    getOptions(): RequestOptions {
+    getOptions() {
         const token = this.authService.loginUserInfo.getValue().token;
         const useName = this.authService.loginUserInfo.getValue().userName;
-        const headers = new Headers(
+        const headers = new HttpHeaders(
             {
                 'Authorization': 'Basic ' + token,
                 'user-name': useName,
                 'Content-Type': 'application/json'
             });
-        return new RequestOptions({headers: headers});
+        return { headers: headers };
     }
 
 }
