@@ -1,80 +1,69 @@
-import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs';
-import {Http, Headers, RequestOptions, Response} from '@angular/http';
-import {User, LoginResponse} from '../commons/models/common-data-models';
+
+import { throwError as observableThrowError, Observable } from 'rxjs';
+import { Injectable } from '@angular/core';
+import { HttpClient, HttpHeaders, HttpBackend } from '@angular/common/http';
+import { map, catchError } from 'rxjs/operators';
+import { User, LoginResponse } from '../commons/models/common-data-models';
 
 
 @Injectable()
 export class LoginRemoteDataService {
 
-  private apiContext: string = 'api';
-  private headers: Headers = new Headers({'Content-Type': 'application/json'});
-  private options: RequestOptions = new RequestOptions({headers: this.headers});
+    private httpBasicClient: HttpClient
+    private apiContext: string = 'api';
+    private apiEndpoints = {
+        login: this.apiContext + '/authentication/login',
+        logout: this.apiContext + '/authentication/logout',
+        getUserDetails: this.apiContext + '/authentication/userdetails'
+    };
 
-  private apiEndpoints: Object = {
-    login: this.apiContext + '/authentication/login',
-    logout: this.apiContext + '/authentication/logout',
-    getUserDetails: this.apiContext + '/authentication/userdetails'
-  };
-
-  constructor(private http: Http) {
-  }
-
-  /**
-   * Login service api
-   * @param data
-   * @returns {Observable<User>}
-   */
-  login(data: User) {
-    return this.http.get(this.apiEndpoints['login'], this.getOptions(data))
-        .map((response: Response) => {
-          return {
-            success: true,
-            message: 'Login Successful',
-            payload: response.json()
-          };
-        })
-        .catch((error: Response) => Observable.throw({
-          success: false,
-          message: 'Login Error',
-          error: error
-        }));
-  }
-
-  /**
-   * Logout Service api
-   * @param userId
-   * @returns {Observable<boolean>}
-   */
-  logout(userId: string): Observable<boolean> {
-    return this.http.get(this.apiEndpoints['logout'] + '/' + userId)
-      .map((response: Response) => response.json().success)
-      .catch((error: Response) => Observable.throw(error.json().message))
-  }
-
-  getUserDetails(userName: string) {
-    // console.log('hit in the rate remote data service to get currency');
-    return this.http.get(this.apiEndpoints['getUserDetails'] + '/' + userName)
-        .map((response: Response) => {
-          const result = response.json();
-          return result;
-        })
-        .catch((error: Response) => Observable.throw(error.json().message()));
-  }
+    constructor(private http: HttpClient, handler: HttpBackend) {
+        this.httpBasicClient = new HttpClient(handler);
+    }
 
     /**
-     * generate request headers
-     * @returns {RequestOptions}
+     * Login service api
+     * @param data
+     * @returns {Observable<User>}
      */
-    getOptions(data: User): RequestOptions {
-        const token = btoa(data.userName + ':' + data.password);
-        const headers = new Headers(
-            {
-                'Authorization': 'Basic ' + token,
-                'user-name': data.userName,
-                'Content-Type': 'application/json'
-            });
-        return new RequestOptions({headers: headers});
+    login(data: User){
+        const httpOption = {
+            headers: new HttpHeaders({
+                'Authorization': 'Basic ' + btoa(`${data.userName}:${data.userName}`),
+                'Content-Type': 'application/json',
+                'user-name': data.userName
+            })
+        };
+
+        return this.httpBasicClient.get(this.apiEndpoints.login, httpOption).pipe(
+            map((response) => {
+                return {
+                  success: true,
+                  message: 'Login Successful',
+                  payload: response
+                };
+              }),
+              catchError((error: Response) => Observable.throw({
+                success: false,
+                message: 'Login Error',
+                error: error
+              }))
+        )
+    }
+
+    /**
+     * Logout Service api
+     * @param userId
+     * @returns {Observable<boolean>}
+     */
+    logout(userId: string): Observable<boolean> {
+        return this.httpBasicClient.get(this.apiEndpoints.logout + '/' + userId).pipe(
+            map((response: Response) => true), catchError((error: Response) => observableThrowError(error['message']))
+        )
+    }
+
+    getUserDetails(userName: string) {
+        return this.http.get(this.apiEndpoints.getUserDetails + '/' + userName)
     }
 
 }
