@@ -1,10 +1,12 @@
+
+import { throwError as observableThrowError, Observable } from 'rxjs';
 /**
  * Created by sahanK on 08/02/17.
  */
-import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs';
-import {Http, Headers, RequestOptions, Response} from '@angular/http';
-import {AuthenticationService} from '../commons/services/authentication.service';
+import { Injectable } from '@angular/core';
+import { AuthenticationService } from '../commons/services/authentication.service';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { map, catchError } from 'rxjs/operators';
 
 
 @Injectable()
@@ -24,7 +26,7 @@ export class QuotacapRemoteDataService {
 
     };
 
-    constructor(private http: Http, private _authenticationService: AuthenticationService) {
+    constructor(private http: HttpClient, private _authenticationService: AuthenticationService) {
         this.loginInfo = this._authenticationService.loginUserInfo.getValue();
 
     }
@@ -35,18 +37,20 @@ export class QuotacapRemoteDataService {
      */
     getSubscribers(operatorName: string) {
         return this.http.get(this.apiEndpoints['subscribers'] + operatorName, this.getOptions())
-            .map((response: Response) => {
-                return {
-                    success: true,
-                    message: 'Subscribers Loaded Successfully',
-                    payload: response.json()
-                };
-            })
-            .catch((error: Response) => Observable.throw({
-                success: false,
-                message: 'Error Loading Subscribers',
-                error: error
-            }));
+            .pipe(
+                map((response) => {
+                    return {
+                        success: true,
+                        message: 'Subscribers Loaded Successfully',
+                        payload: response
+                    };
+                }),
+                catchError((error: Response) => observableThrowError({
+                    success: false,
+                    message: 'Error Loading Subscribers',
+                    error: error
+                }))
+            )
     }
 
     /**
@@ -55,18 +59,20 @@ export class QuotacapRemoteDataService {
      */
     getOperatorOfsubscriber(subscriberID: string) {
         return this.http.get(this.apiEndpoints['operators'] + subscriberID, this.getOptions())
-            .map((response: Response) => {
-                return {
-                    success: true,
-                    message: 'Operators Loaded Successfully',
-                    payload: response.json()
-                };
-            })
-            .catch((error: Response) => Observable.throw({
-                success: false,
-                message: 'Error Loading Operators',
-                error: error
-            }));
+            .pipe(
+                map((response) => {
+                    return {
+                        success: true,
+                        message: 'Operators Loaded Successfully',
+                        payload: response
+                    };
+                }),
+                catchError((error: Response) => observableThrowError({
+                    success: false,
+                    message: 'Error Loading Operators',
+                    error: error
+                }))
+            )
     }
 
     /**
@@ -76,34 +82,38 @@ export class QuotacapRemoteDataService {
      */
     getQuotaLimitInfo(data) {
         return this.http.post(this.apiEndpoints['quotaLimitInfo'], data, this.getOptions())
-            .map((response: Response) => {
-                return {
-                    success: true,
-                    message: 'Quota Limit Info Loaded Successfully',
-                    payload: response.json()
-                };
-            })
-            .catch((error: Response) => Observable.throw({
-                success: false,
-                message: 'Error Loading Quota Limit Info',
-                error: error
-            }));
+            .pipe(
+                map((response) => {
+                    return {
+                        success: true,
+                        message: 'Quota Limit Info Loaded Successfully',
+                        payload: response
+                    };
+                }),
+                catchError((error: Response) => observableThrowError({
+                    success: false,
+                    message: 'Error Loading Quota Limit Info',
+                    error: error
+                }))
+            )
     }
 
     getValidityPeriod(data) {
         return this.http.post(this.apiEndpoints['validityPeriod'], data, this.getOptions())
-            .map((response: Response) => {
-                return {
-                    success: true,
-                    message: 'Period Validation Success',
-                    payload: response.json()
-                };
-            })
-            .catch((error: Response) => Observable.throw({
-                success: false,
-                message: 'Error While Validating Time Period',
-                error: error
-            }));
+            .pipe(
+                map((response) => {
+                    return {
+                        success: true,
+                        message: 'Period Validation Success',
+                        payload: response
+                    };
+                }),
+                catchError((error: Response) => observableThrowError({
+                    success: false,
+                    message: 'Error While Validating Time Period',
+                    error: error
+                }))
+            )
     }
 
     /**
@@ -116,30 +126,32 @@ export class QuotacapRemoteDataService {
      */
     addNewQuotaLimit(data) {
         return this.http.post(this.apiEndpoints['addNewQuotaLimit'], data, this.getOptions())
-            .map((response: Response) => {
-                return {
-                    success: true,
-                    message: 'New Quota Value Added Successfully',
-                    payload: response.json()
-                };
-            })
-            .catch((error: Response) => Observable.throw({
-                success: false,
-                message: 'Error While Adding New Quota Limit',
-                error: error
-            }));
+            .pipe(
+                map((response) => {
+                    return {
+                        success: true,
+                        message: 'New Quota Value Added Successfully',
+                        payload: response
+                    };
+                }),
+                catchError((error: Response) => observableThrowError({
+                    success: false,
+                    message: 'Error While Adding New Quota Limit',
+                    error: error
+                }))
+            )
     }
 
-    getOptions(): RequestOptions {
+    getOptions() {
         const token = this._authenticationService.loginUserInfo.getValue().token;
         const useName = this._authenticationService.loginUserInfo.getValue().userName;
-        const headers = new Headers(
+        const headers = new HttpHeaders(
             {
                 'Authorization': 'Basic ' + token,
                 'user-name': useName,
                 'Content-Type': 'application/json'
             });
-        return new RequestOptions({headers: headers});
+        return { headers: headers };
     }
 
 
