@@ -1,6 +1,6 @@
+
+import {throwError as observableThrowError, Subject, BehaviorSubject, Observable} from 'rxjs';
 import {Injectable, Inject} from '@angular/core';
-import {Headers, RequestOptions, Http, Response} from "@angular/http";
-import {Subject, BehaviorSubject, Observable} from "rxjs";
 import {MessageService} from "../commons/services/message.service";
 import {SlimLoadingBarService} from "ng2-slim-loading-bar";
 import {
@@ -8,6 +8,8 @@ import {
     Application, ApplicationHistory, AppHistoryResponse, SubscriptionHistoryResponse, SubscriptionHistoryFilter
 } from "../commons/models/reporing-data-models";
 import {AuthenticationService} from '../commons/services/authentication.service';
+import { HttpHeaders, HttpClient } from '@angular/common/http';
+import { catchError } from 'rxjs/operators';
 
 @Injectable()
 export class ReportingRemoteDataService {
@@ -16,13 +18,13 @@ export class ReportingRemoteDataService {
      * Subscribers stream
      * @type {BehaviorSubject<string[]>}
      */
-    public SubscribersProvider: Subject<string[]> = new BehaviorSubject<string[]>([]);
+    public SubscribersProvider: Subject<any> = new BehaviorSubject<string[]>([]);
 
     /**
      * Operators Stream
      * @type {BehaviorSubject<string[]>}
      */
-    public OperatorsProvider: Subject<string[]> = new BehaviorSubject<string[]>([]);
+    public OperatorsProvider: Subject<any> = new BehaviorSubject<string[]>([]);
 
     /**
      * Applications Stream
@@ -38,9 +40,9 @@ export class ReportingRemoteDataService {
      */
     public ApprovalHistoryProvider: Subject<AppHistoryResponse> = new BehaviorSubject<AppHistoryResponse>(null);
 
-    private headers: Headers = new Headers({'Content-Type': 'application/json'});
+    private headers: HttpHeaders = new HttpHeaders({'Content-Type': 'application/json'});
 
-    private options: RequestOptions = new RequestOptions({headers: this.headers});
+    private options = {headers: this.headers};
 
     private url = new URL(window.location.href);
     private apiContext = this.url.protocol + '//' + this.url.host + '/workflow-service/workflow/';
@@ -54,17 +56,16 @@ export class ReportingRemoteDataService {
         applicationHistory: this.apiContext + 'applications/history'
     };
 
-    constructor(private http: Http,
+    constructor(private http: HttpClient,
                 private message: MessageService,
                 private slimLoadingBarService: SlimLoadingBarService,
                 private authService: AuthenticationService) {
     }
 
     getApplicationDetail(id: number, callback: Function) {
-        this.http.get(this.apiEndpoints['approvalHistory'] + '/' + id, this.getOptions())
-            .map((response: Response) => response.json())
+        this.http.get<ApplicationHistory[]>(this.apiEndpoints['approvalHistory'] + '/' + id, this.getOptions())
             .subscribe(
-                (applications: ApplicationHistory[]) => {
+                (applications) => {
                     this.ApplicationDetailProvider.next(applications);
                     callback(applications, true);
                 },
@@ -78,7 +79,6 @@ export class ReportingRemoteDataService {
 
     getSubscriptionDetail(id: number,opId:string,apiid:string, callback: Function) {
         this.http.get(this.apiEndpoints['approvalHistory'] + '/' + id + '/operators/' + opId + '/apis/' + apiid + '/start/0/size/50'  , this.getOptions())
-            .map((response: Response) => response.json())
             .subscribe(
                 (applications: ApplicationHistory[]) => {
                     this.ApplicationDetailProvider.next(applications);
@@ -94,7 +94,6 @@ export class ReportingRemoteDataService {
     getSubscribers() {
         this.slimLoadingBarService.start();
         this.http.get(this.apiEndpoints['subscribers'], this.getOptions())
-            .map((response: Response) => response.json())
             .subscribe(
                 (subscribers) => {
                     this.SubscribersProvider.next(subscribers)
@@ -112,7 +111,6 @@ export class ReportingRemoteDataService {
     getOperators() {
         this.slimLoadingBarService.start();
         this.http.get(this.apiEndpoints['operators'], this.getOptions())
-            .map((response: Response) => response.json())
             .subscribe(
                 (operators) => {
                     this.OperatorsProvider.next(operators)
@@ -131,7 +129,6 @@ export class ReportingRemoteDataService {
         if (!!subscriber) {
             this.slimLoadingBarService.start();
             this.http.get(this.apiEndpoints['applications'] + '/' + subscriber, this.getOptions())
-                .map((response: Response) => response.json())
                 .subscribe(
                     (applications: Application[]) => {
                         this.ApplicationsProvider.next(applications)
@@ -164,19 +161,20 @@ export class ReportingRemoteDataService {
             + '?start=' + historyFilter.offset + '&filterBy=' + historyFilter.filterString;
 
         this.http.get(endPoint, this.getOptions())
-            .map((response: Response) => response.json())
-            .catch((error: Response) => Observable.throw({
-                success: false,
-                message: 'Error Loading Application History List',
-                error: error
-            }))
+            .pipe(
+                catchError((error: Response) => observableThrowError({
+                    success: false,
+                    message: 'Error Loading Application History List',
+                    error: error
+                }))
+            )
             .subscribe(
                 data => {
-                    if (data.success) {
-                        this.ApprovalHistoryProvider.next(data.payload);
+                    if (data['success']) {
+                        this.ApprovalHistoryProvider.next(data['payload']);
                         this.slimLoadingBarService.complete();
                     } else {
-                        this.message.error(data.message);
+                        this.message.error(data['message']);
                         this.slimLoadingBarService.stop();
                     }
                 },
@@ -187,16 +185,16 @@ export class ReportingRemoteDataService {
             );
     }
 
-    getOptions(): RequestOptions {
+    getOptions() {
         const token = this.authService.loginUserInfo.getValue().token;
         const useName = this.authService.loginUserInfo.getValue().userName;
-        const headers = new Headers(
+        const headers = new HttpHeaders(
             {
                 'Authorization': 'Basic ' + token,
                 'user-name': useName,
                 'Content-Type': 'application/json'
             });
-        return new RequestOptions({headers: headers});
+        return {headers: headers};
     }
 
     getDeploymentType(): Promise<any> {
